Remove Constraint; replace SimpleCriterion by BiomeCriterion and StructureCriterion
